+++ stroom-pipeline/src/main/java/stroom/pipeline/refdata/ReferenceDataResource.java
@@ -1,6 +1,7 @@
 package stroom.pipeline.refdata;
 
 import stroom.pipeline.refdata.store.RefStoreEntry;
+import stroom.pipeline.refdata.store.RefStreamProcessingInfo;
 import stroom.util.shared.ResourcePaths;
 import stroom.util.shared.RestResource;
 
@@ -29,6 +30,7 @@ public interface ReferenceDataResource extends RestResource {
     String BASE_PATH = "/refData" + ResourcePaths.V1;
 
     String ENTRIES_SUB_PATH = "/entries";
+    String REF_STREAM_INFO_SUB_PATH = "/refStreamInfo";
     String LOOKUP_SUB_PATH = "/lookup";
     String PURGE_SUB_PATH = "/purge";
 
@@ -39,7 +41,20 @@ public interface ReferenceDataResource extends RestResource {
             description = "This is primarily intended  for small scale debugging in non-production environments. If " +
                     "no limit is set a default limit is applied else the results will be limited to limit entries.",
             operationId = "getReferenceStoreEntries")
-    List<RefStoreEntry> entries(@QueryParam("limit") final Integer limit);
+    List<RefStoreEntry> entries(@QueryParam("limit") final Integer limit,
+                                @QueryParam("refStreamId") final Long refStreamId,
+                                @QueryParam("mapName") final String mapName);
+
+    @GET
+    @Path(REF_STREAM_INFO_SUB_PATH)
+    @Operation(
+            summary = "List processing info entries for all ref streams",
+            description = "This is primarily intended  for small scale debugging in non-production environments. If " +
+                    "no limit is set a default limit is applied else the results will be limited to limit entries.",
+            operationId = "getReferenceStreamProcessingInfoEntries")
+    List<RefStreamProcessingInfo> refStreamInfo(@QueryParam("limit") final Integer limit,
+                                                @QueryParam("refStreamId") final Long refStreamId,
+                                                @QueryParam("mapName") final String mapName);
 
     @POST
     @Path(LOOKUP_SUB_PATH)